Move enums out of OHttpCryptoProvider
Motivation:

It makes the code cleaner when we move the used enums out of OHttpCryptoProvider

Modifications:

Move enums out of OHttpCryptoProvider and rename Mode to HPKEMode

Result:

Cleanup
