+++ codec-ohttp/src/main/java/io/netty/incubator/codec/ohttp/OHttpKey.java
@@ -17,14 +17,14 @@ package io.netty.incubator.codec.ohttp;
 
 import io.netty.incubator.codec.hpke.AsymmetricCipherKeyPair;
 import io.netty.incubator.codec.hpke.CryptoException;
-import io.netty.incubator.codec.hpke.OHttpCryptoProvider.AEAD;
-import io.netty.incubator.codec.hpke.OHttpCryptoProvider.KDF;
+import io.netty.incubator.codec.hpke.AEAD;
+import io.netty.incubator.codec.hpke.KDF;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static io.netty.incubator.codec.hpke.OHttpCryptoProvider.KEM;
+import io.netty.incubator.codec.hpke.KEM;
 import static java.util.Objects.requireNonNull;
 
 public abstract class OHttpKey {
