+++ codec-ohttp/src/main/java/io/netty/incubator/codec/ohttp/OHttpServerPublicKeys.java
@@ -28,9 +28,11 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import static io.netty.incubator.codec.hpke.OHttpCryptoProvider.AEAD;
-import static io.netty.incubator.codec.hpke.OHttpCryptoProvider.KDF;
-import static io.netty.incubator.codec.hpke.OHttpCryptoProvider.KEM;
+import io.netty.incubator.codec.hpke.AEAD;
+
+import io.netty.incubator.codec.hpke.KDF;
+
+import io.netty.incubator.codec.hpke.KEM;
 import static java.util.Objects.requireNonNull;
 
 /**
